📊 multer.js
one file for both image and docs
